back/snack/requirements.txt
asgiref==3.8.1
boto3==1.37.11
botocore==1.37.11
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==3.2.3
django-redis==5.4.0
django-storages==1.14.5
djangorestframework==3.15.2
idna==3.10
jmespath==1.0.1
mysqlclient==2.2.7
numpy==2.2.4
pandas==2.2.3
pycryptodome==3.22.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0